+++ packages/client/src/components/common/DataGrid/DataGrid.tsx
@@ -19,12 +19,11 @@ export const DataGrid = <TData extends Record<string, unknown>>({
   state,
   ...props
 }: React.ComponentProps<typeof MantineReactTable<TData>>) => {
-  const { paginationState, dispatchPaginationState } = usePagination();
+  const { paginationState, dispatchPaginationState, pageSizeOptions } = usePagination();
   const { setSearchQuery } = useSearch();
 
   const dataGridState: MRT_TableState<TData> = React.useMemo(
-    // () => _.defaultsDeep(_.cloneDeep(state), { pagination: paginationState }),
-    () => ({ ...state, pagination: paginationState } as any),
+    () => _.defaultsDeep(state, { pagination: paginationState }),
     [state, paginationState],
   );
 
@@ -47,6 +46,7 @@ export const DataGrid = <TData extends Record<string, unknown>>({
           <IconDatabaseOff stroke={1.5} size="1.25rem" />
         </Group>
       )}
+      mantinePaginationProps={{ rowsPerPageOptions: pageSizeOptions.map(String) }}
       enableToolbarInternalActions={enableToolbarInternalActions}
       onPaginationChange={onPaginationChange}
       onGlobalFilterChange={setSearchQuery}