+++ packages/client/src/pages/CoursesPage/CoursesPageAdmin/CoursesPageAdmin.tsx
@@ -3,6 +3,7 @@ import React from 'react';
 import { Button, Menu } from '@mantine/core';
 import { notifications } from '@mantine/notifications';
 import {
+  CourseWhere,
   GetCoursesDocument,
   SortDirection,
   useCreateCourseMutation,
@@ -19,12 +20,15 @@ import {
   IconTrash,
   IconX,
 } from '@tabler/icons-react';
+import _ from 'lodash';
+import { MRT_ColumnFiltersState } from 'mantine-react-table';
 import { Link, generatePath } from 'react-router-dom';
 
 import { ContentPageLayout, CreateUpdateCourseModal, DataGrid } from 'src/components';
 import { CreateUpdateCourseModalArgs } from 'src/components/modals/CreateUpdateCourse/types';
-import { ModalId, RouteEntityType, appRoutes } from 'src/constants';
+import { ModalId, RouteEntityType, TOPIC_TAGS, appRoutes } from 'src/constants';
 import {
+  useDebounce,
   useModal,
   usePagination,
   usePaginationQueryOptions,
@@ -35,6 +39,8 @@ import { PaginationActionType } from 'src/providers';
 import { COURSES_TABLE_COLUMNS } from '../constants';
 
 export const CoursesPageAdmin: React.FC = () => {
+  const [columnFilters, setColumnFilters] = React.useState<MRT_ColumnFiltersState>([]);
+
   const { openModal, closeModal, switchClosability } = useModal<CreateUpdateCourseModalArgs>(
     ModalId.CreateUpdateCourseModal,
   );
@@ -42,13 +48,6 @@ export const CoursesPageAdmin: React.FC = () => {
   const { paginationState, dispatchPaginationState } = usePagination();
   const searchOptions = useSearchQueryOptions(['title', 'description']);
 
-  const { data, loading: loadingCourses } = useGetCoursesQuery({
-    variables: {
-      where: searchOptions,
-      options: { ...paginationOptions, sort: [{ createdAt: SortDirection.Desc }] },
-    },
-    notifyOnNetworkStatusChange: true,
-  });
   const [createCourse, { loading: creatingCourse }] = useCreateCourseMutation();
   const [updateCourse, { loading: updatingCourse }] = useUpdateCourseByIdMutation();
   const [deleteCourse] = useDeleteCourseByIdMutation();
@@ -56,14 +55,46 @@ export const CoursesPageAdmin: React.FC = () => {
   const columns = React.useMemo(
     () =>
       COURSES_TABLE_COLUMNS({
+        // FIXME: Get from global state
         languages: ['ru', 'en', 'fr', 'es', 'de'],
         platforms: ['Stepik', 'Coursera'],
-        difficulties: ['beginner', 'intermediate', 'advanced'],
-        topics: [],
-        setFiltersState: () => void 0,
+        difficulties: ['BEGINNER', 'INTERMEDIATE', 'ADVANCED'],
+        topics: TOPIC_TAGS,
+        setFiltersState: setColumnFilters,
       }).filter(col => !['id'].includes(col.accessorKey as string)),
     [],
   );
+
+  const coursesFilters = React.useMemo((): CourseWhere => {
+    return columnFilters.reduce((acc, colFilter) => {
+      if (typeof colFilter.value === 'string') {
+        return _.set(acc, `${colFilter.id}_CONTAINS`, colFilter.value);
+      }
+
+      if (Array.isArray(colFilter.value) && !_.isEmpty(colFilter.value.filter(Boolean))) {
+        return _.set(acc, `${colFilter.id}_IN`, colFilter.value);
+      }
+
+      if (typeof colFilter.value === 'object') {
+        const val = colFilter.value as { min?: number; max?: number };
+        _.set(acc, `${colFilter.id}_GTE`, val.min);
+        _.set(acc, `${colFilter.id}_LTE`, val.max);
+        return acc;
+      }
+
+      return acc;
+    }, {} as CourseWhere);
+  }, [columnFilters]);
+
+  const debouncedCoursesFilters = useDebounce(coursesFilters, 1000);
+
+  const { data, loading: loadingCourses } = useGetCoursesQuery({
+    variables: {
+      where: { ...searchOptions, ...debouncedCoursesFilters },
+      options: { ...paginationOptions, sort: [{ createdAt: SortDirection.Desc }] },
+    },
+    notifyOnNetworkStatusChange: true,
+  });
 
   const handleCreateFormSubmit: CreateUpdateCourseModalArgs['onSubmit'] = async ({
     platformId,
@@ -192,7 +223,16 @@ export const CoursesPageAdmin: React.FC = () => {
 
     // TODO: Confirmation modal
     if (window.confirm('Are you sure you want to delete the course?')) {
+      // FIXME: Update cache
       await deleteCourse({ variables: { id }, refetchQueries: [GetCoursesDocument] });
+      notifications.show({
+        title: 'Success',
+        message: 'Course has been deleted',
+        color: 'green',
+        withCloseButton: true,
+        icon: <IconCheck />,
+        withBorder: true,
+      });
     }
   };
 
@@ -220,12 +260,14 @@ export const CoursesPageAdmin: React.FC = () => {
         }}
         columns={columns}
         data={courses ?? []}
-        state={{ isLoading: loadingCourses }}
+        state={{ isLoading: loadingCourses, columnFilters }}
         enableColumnOrdering
         rowCount={paginationState.count}
         initialState={{
           showGlobalFilter: true,
         }}
+        manualFiltering
+        onColumnFiltersChange={setColumnFilters}
         positionGlobalFilter="left"
         enableRowActions
         renderRowActionMenuItems={({
@@ -235,7 +277,7 @@ export const CoursesPageAdmin: React.FC = () => {
         }) => (
           <React.Fragment>
             <Menu.Item
-              icon={<IconListDetails />}
+              icon={<IconListDetails size="1rem" />}
               component={Link}
               to={generatePath(appRoutes.courses.details, {
                 [RouteEntityType.course]: id,
@@ -244,7 +286,7 @@ export const CoursesPageAdmin: React.FC = () => {
               View details
             </Menu.Item>
             <Menu.Item
-              icon={<IconEdit />}
+              icon={<IconEdit size="1rem" />}
               onClick={() =>
                 openModal(ModalId.CreateUpdateCourseModal, {
                   onSubmit: formData => handleUpdateFormSubmit(formData, id),
@@ -266,7 +308,11 @@ export const CoursesPageAdmin: React.FC = () => {
             >
               Edit course
             </Menu.Item>
-            <Menu.Item icon={<IconTrash />} onClick={async () => await handleCourseDelete(id)}>
+            <Menu.Item
+              icon={<IconTrash size="1rem" />}
+              color="red"
+              onClick={async () => await handleCourseDelete(id)}
+            >
               Delete course
             </Menu.Item>
           </React.Fragment>