+++ packages/parser/src/common/parsers/coursera/utils.ts
@@ -1,15 +1,15 @@
-import { CourseDifficulty } from '@shared/graphql';
+import { Difficulty } from '@shared/graphql';
 import type { Nullable } from '@shared/types';
 
-export const mapCourseraDifficulty = (difficulty: Nullable<string>): Nullable<CourseDifficulty> => {
+export const mapCourseraDifficulty = (difficulty: Nullable<string>): Nullable<Difficulty> => {
   if (difficulty?.match(/.*beginner.*/gi)) {
-    return CourseDifficulty.Beginner;
+    return Difficulty.Beginner;
   }
   if (difficulty?.match(/.*intermediate.*/gi)) {
-    return CourseDifficulty.Intermediate;
+    return Difficulty.Intermediate;
   }
   if (difficulty?.match(/.*advanced.*/gi)) {
-    return CourseDifficulty.Advanced;
+    return Difficulty.Advanced;
   }
 
   return null;